#17: Refactor to remove unnecessary fragment.
