+++ app/src/main/java/com/example/concordia_campus_guide/Activities/SearchActivity.java
@@ -5,22 +5,20 @@ import android.location.Location;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
-import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.EditText;
 import android.widget.ImageButton;
+import android.widget.ImageView;
 import android.widget.ListView;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.constraintlayout.widget.ConstraintLayout;
-import androidx.constraintlayout.widget.Group;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProviders;
 
 import com.example.concordia_campus_guide.Adapters.PlaceToSearchResultAdapter;
-import com.example.concordia_campus_guide.Fragments.CalendarFragment.CalendarFragment;
+import com.example.concordia_campus_guide.Models.Helpers.CalendarViewModel;
 import com.example.concordia_campus_guide.Global.SelectingToFromState;
 import com.example.concordia_campus_guide.Models.Building;
 import com.example.concordia_campus_guide.Models.CalendarEvent;
@@ -30,15 +28,15 @@ import com.example.concordia_campus_guide.Models.Place;
 import com.example.concordia_campus_guide.Models.RoomModel;
 import com.example.concordia_campus_guide.R;
 
-import java.util.zip.Inflater;
-
 public class SearchActivity extends AppCompatActivity {
 
-//    ListView searchResults;
+    ListView searchResults;
     EditText searchText;
     View nextClassRow;
-    Fragment nextClassText;
+    TextView nextClassText;
+    ImageView nextClassArrow;
     SearchActivityViewModel mViewModel;
+    CalendarViewModel calendarViewModel;
 
     private PlaceToSearchResultAdapter adapter;
 
@@ -49,15 +47,18 @@ public class SearchActivity extends AppCompatActivity {
         //setting up activity
         setContentView(R.layout.search_activity);
         setBackButtonOnClick();
-//        searchResults = (ListView) findViewById(R.id.searchResults);
+        searchResults = (ListView) findViewById(R.id.searchResults);
         searchText = (EditText) findViewById(R.id.searchText);
         mViewModel = ViewModelProviders.of(this).get(SearchActivityViewModel.class);
+        nextClassText = (TextView) findViewById(R.id.next_class_text);
+        nextClassArrow = (ImageView) findViewById(R.id.next_class_arrow);
         nextClassRow = (View) findViewById(R.id.view_container);
-        nextClassText = (Fragment) getSupportFragmentManager().findFragmentById(R.id.next_class_text);
+        calendarViewModel = ViewModelProviders.of(this).get(CalendarViewModel.class);
+
 
         //android adapter for list view
         adapter = new PlaceToSearchResultAdapter(this, R.layout.list_item_layout, mViewModel.getAllPlaces());
-//        searchResults.setAdapter(adapter);
+        searchResults.setAdapter(adapter);
 
         //search results filtering according to search text
         searchText.addTextChangedListener(new TextWatcher(){
@@ -72,25 +73,33 @@ public class SearchActivity extends AppCompatActivity {
             public void afterTextChanged(Editable editable){ }
         });
 
-//        searchResults.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-//            @Override
-//            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-//                Place place = adapter.getItem(position);
-//                openRoutesPage(place);
-//            }
-//        });
-
-        nextClassRow.setOnClickListener(new View.OnClickListener() {
+        searchResults.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
-            public void onClick(View v) {
-                openRoutesPage(new Place() {
-                    @Override
-                    public String getDisplayName() {
-                        return nextClassText.toString();
-                    }
-                });
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                Place place = adapter.getItem(position);
+                openRoutesPage(place);
             }
         });
+
+        final CalendarEvent calendarEvent = calendarViewModel.getEvent(getApplicationContext(), this);
+
+        nextClassText.setText(calendarViewModel.getNextClassString(calendarEvent));
+
+        if(nextClassText.getText() != "No classes today"){
+            nextClassArrow.setVisibility(View.VISIBLE);
+            nextClassRow.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    openRoutesPage(new Place() {
+                        @Override
+                        public String getDisplayName() {
+                            return calendarEvent.getLocation();
+                        }
+                    });
+                }
+            });
+        }
+
     }
 
     private void openRoutesPage(Place place){
