--- app/src/main/java/com/example/concordia_campus_guide/Fragments/CalendarFragment/CalendarViewModel.java
@@ -1,59 +1,0 @@
-package com.example.concordia_campus_guide.Fragments.CalendarFragment;
-
-import android.content.ContentUris;
-import android.content.Context;
-import android.database.Cursor;
-import android.net.Uri;
-import android.provider.CalendarContract;
-import android.provider.CalendarContract.Instances;
-
-import androidx.lifecycle.ViewModel;
-
-import com.example.concordia_campus_guide.Models.CalendarEvent;
-
-public class CalendarViewModel extends ViewModel {
-
-    // The indices for the projection array above.
-    private static final int PROJECTION_TITLE_INDEX = 0;
-    private static final int PROJECTION_LOCATION_INDEX = 1;
-    private static final int PROJECTION_START_INDEX = 2;
-
-    public static final String[] INSTANCE_PROJECTION = new String[] {
-            Instances.TITLE,
-            Instances.EVENT_LOCATION,
-            Instances.DTSTART
-    };
-
-    public Cursor getCalendarCursor(Context context) {
-        Uri.Builder eventsUriBuilder = CalendarContract.Instances.CONTENT_URI
-                .buildUpon();
-        long startTime = System.currentTimeMillis();
-        long endTime = startTime + 28800000;
-        ContentUris.appendId(eventsUriBuilder, startTime);
-        ContentUris.appendId(eventsUriBuilder, endTime);
-        Uri eventsUri = eventsUriBuilder.build();
-        Cursor cursor = context.getContentResolver().query(
-                eventsUri,
-                INSTANCE_PROJECTION,
-                null,
-                null,
-                Instances.BEGIN + " ASC"
-        );
-
-        return cursor;
-    }
-
-    public CalendarEvent getCalendarEvent(Cursor cursor) {
-
-        while (cursor.moveToNext()) {
-           String eventTitle = cursor.getString(PROJECTION_TITLE_INDEX);
-           String eventLocation = cursor.getString(PROJECTION_LOCATION_INDEX);
-           String eventStart = cursor.getString(PROJECTION_START_INDEX);
-
-           if(eventTitle.contains("Lecture: ")||eventTitle.contains("Tutorial: ") || eventTitle.contains("Lab: ")){
-               return new CalendarEvent(eventTitle, eventLocation, eventStart);
-           }
-        }
-        return null;
-    }
-}
