+++ app/src/main/res/layout/activity_main.xml
@@ -61,13 +61,5 @@
 
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
 
-    <fragment
-        android:id="@+id/fragment2"
-        android:name="com.example.concordia_campus_guide.Fragments.CalendarFragment.CalendarFragment"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
